docs(metrics): split generated benchmark sources
